[DEL] Remove third-party PyTorch convolution
Instead, JAX is used to check correctness of higher-order convolution.
There seems to be a bug for specific choices of padding in the third-
party implementation. There is a test case for which einconv and JAX
agree, while einconv and third-party disagree. This hints that the
bug is likely not in einconv.